ci: build all packages and bootitems in CI
Also moved the "nix flake check" to the CI remote builder
to benefit from caching.

The main idea here is to decouple the build of packages, bootitems,
and other convenience/exports from "nix flake check" but also the
build of the NixOS systems.

This way, it is easier to spot what fails and we can better see which
steps consume how much time.