common(modules,user-env): update ZSH
The SHARE_HISTORY option doesn't really fit my workflow.
The problem is that with multiple terminals, what ever I
do in terminal two, will soon be available in terminal one
when I hit "arrow up" to get to the previous command.

However, typically, I want to go back to the previous
command of that terminal (that ZSH session) and not the
previous system-wide ZSH command.