common(services,flake-prefetch): move script to dedicate package
Unfortunately, on all my machines, the Nix garbage collection removes the
directory that is referenced by the ExecStart line. For some reason I don't
understand yet, Nix doesn't track the Nix store path is in use. This way, we
make it clear to Nix that the file is used as it is in $PATH of the user.